+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 #### Improvements
 * Fix #1880: Remove use of reapers manually doing cascade deletion of resources, leave it upto Kubernetes APIServer
+* Test coverage for ServiceCatalog
 * Fix #1772: Add Javadocs for KubernetesClient class
 
 #### Dependency Upgrade